Controllers completed

+++ src/main/java/by/afanasyeu/avtoxam/service/impl/FavoriteServiceImpl.java
@@ -3,6 +3,7 @@ package by.afanasyeu.avtoxam.service.impl;
 import by.afanasyeu.avtoxam.dao.entities.Favorite;
 import by.afanasyeu.avtoxam.dao.mappers.FavoriteMapper;
 import by.afanasyeu.avtoxam.service.FavoriteService;
+import org.apache.ibatis.annotations.Select;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -21,11 +22,10 @@ public class FavoriteServiceImpl implements FavoriteService{
 
     @Override
     public void insertFavorite(Favorite favorite) {
-        favoriteMapper.insertFavorite(favorite);
-    }
-
-    @Override
-    public void delete(Favorite favorite) {
-        favoriteMapper.delete(favorite);
+        if (favoriteMapper.isExist(favorite)) {
+            favoriteMapper.delete(favorite);
+        } else {
+            favoriteMapper.insertFavorite(favorite);
+        }
     }
 }
